refactor: 공통 패키지 분리
