+++ src/main/java/com/whatpl/domain/member/domain/Member.java
@@ -1,6 +1,6 @@
 package com.whatpl.domain.member.domain;
 
-import com.whatpl.global.common.BaseTimeEntity;
+import com.whatpl.global.common.domain.BaseTimeEntity;
 import com.whatpl.global.common.domain.enums.*;
 import com.whatpl.global.exception.BizException;
 import com.whatpl.global.exception.ErrorCode;
