+++ src/main/java/com/whatpl/domain/member/domain/MemberReference.java
@@ -1,6 +1,6 @@
 package com.whatpl.domain.member.domain;
 
-import com.whatpl.global.common.BaseTimeEntity;
+import com.whatpl.global.common.domain.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
