+++ src/main/java/com/whatpl/domain/member/domain/MemberSkill.java
@@ -1,6 +1,6 @@
 package com.whatpl.domain.member.domain;
 
-import com.whatpl.global.common.BaseTimeEntity;
+import com.whatpl.global.common.domain.BaseTimeEntity;
 import com.whatpl.global.common.domain.enums.Skill;
 import jakarta.persistence.*;
 import lombok.*;
