+++ src/main/java/com/whatpl/domain/member/domain/MemberSubject.java
@@ -1,6 +1,6 @@
 package com.whatpl.domain.member.domain;
 
-import com.whatpl.global.common.BaseTimeEntity;
+import com.whatpl.global.common.domain.BaseTimeEntity;
 import com.whatpl.global.common.domain.enums.Subject;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
