+++ src/main/java/com/whatpl/domain/project/domain/ProjectLike.java
@@ -1,6 +1,6 @@
 package com.whatpl.domain.project.domain;
 
-import com.whatpl.global.common.BaseTimeEntity;
+import com.whatpl.global.common.domain.BaseTimeEntity;
 import com.whatpl.domain.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
