+++ src/main/java/com/whatpl/domain/project/domain/RecruitJob.java
@@ -1,6 +1,6 @@
 package com.whatpl.domain.project.domain;
 
-import com.whatpl.global.common.BaseTimeEntity;
+import com.whatpl.global.common.domain.BaseTimeEntity;
 import com.whatpl.global.common.domain.enums.Job;
 import com.whatpl.global.exception.BizException;
 import com.whatpl.global.exception.ErrorCode;
