+++ src/main/java/com/whatpl/global/common/controller/HealthCheckController.java
@@ -1,4 +1,4 @@
-package com.whatpl.global.common;
+package com.whatpl.global.common.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
