+++ src/main/java/com/whatpl/global/common/properties/ImageProperties.java
@@ -1,4 +1,4 @@
-package com.whatpl.global.properties;
+package com.whatpl.global.common.properties;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
