+++ src/test/java/com/whatpl/BaseSecurityWebMvcTest.java
@@ -6,8 +6,8 @@ import com.whatpl.domain.chat.service.ChatMessageService;
 import com.whatpl.domain.chat.service.ChatRoomService;
 import com.whatpl.domain.project.service.*;
 import com.whatpl.global.config.SecurityConfig;
-import com.whatpl.global.properties.ImageProperties;
-import com.whatpl.global.properties.JwtProperties;
+import com.whatpl.global.common.properties.ImageProperties;
+import com.whatpl.global.common.properties.JwtProperties;
 import com.whatpl.global.jwt.service.JwtService;
 import com.whatpl.external.upload.FileUploader;
 import com.whatpl.domain.member.service.MemberLoginService;
